Added support for rollback

+++ patch-cluster-node.sh
@@ -5,11 +5,21 @@
 # To fully patch a cluster, this script must be run on EVERY node to update the local filesystem. On ONE NODE ONLY
 # the -a switch must be specified to patch the HDFS contents.
 
+# Parameters
 APPLY_HDFS_PATCH=0
 HDFS_USER=
 TARGET_RELEASE=
 DIR_PREFIX=
 HDFS_DIR_PREFIX=
+ROLLBACK=0
+
+# Constants
+export MATCHED_JAR_FILE_NAME=hadoop-azure
+GITHUB_API_ROOT_URI=https://api.github.com/repos/jamesbak/abfs_backport
+BACKUP_SUFFIX=".original"
+JAR_EXTENSION=".jar"
+#
+JAR_FIND_SUFFIX=""
 
 while getopts ":a?hu:t:p:P:" options
 do
@@ -29,8 +39,11 @@ do
         P)
             HDFS_DIR_PREFIX=${OPTARG}
             ;;
+        R)
+            ROLLBACK=1
+            ;;
         *|?|h)
-            echo "Usage: $0 [-a] [-u HDFS_USER] [-t TARGET_VERSION] [-p DIRECTORY_PREFIX] [-P HDFS_DIRECTORY_PREFIX] [-?]"
+            echo "Usage: $0 [-a] [-u HDFS_USER] [-t TARGET_VERSION] [-p DIRECTORY_PREFIX] [-P HDFS_DIRECTORY_PREFIX] [-R] [-?]"
             echo ""
             echo "Where:"
             echo "  -a              Update HDFS contents. This switch should only be specified for ONE node in a cluster patch."
@@ -41,6 +54,8 @@ do
             echo "                  Specifies a prefix that is specific to the Hadoop distro & version to search for files to patch."
             echo "  -P HDFS_DIRECTORY_PREFIX "
             echo "                  Specifies a prefix that is specific to the Hadoop distro & version to search on HDFS for files to patch."
+            echo "  -R              Rollback installation. Restores previously backed up versions of hadoop-azure jar file. Rollback for HDFS "
+            echo "                  should follow same model as deployment."
             exit 1
             ;;
     esac
@@ -53,13 +68,28 @@ if [ $? -ne 0 ]; then
     exit 1
 fi
 
-GITHUB_API_ROOT_URI=https://api.github.com/repos/jamesbak/abfs_backport
+# Confirm rollback
+if [ $ROLLBACK -gt 0 ]; then
+
+    echo "***************** NOTICE ****************************"
+    echo "This script will rollback previously applied changes."
+    echo "Multiple patches and rollbacks are NOT idempotent. Rolling back after applying multiple patches "
+    echo "may result in an unusable system."
+    echo "Only rollback if you are confident there is only one saved original version (*.jar.original)."
+    read -r -p "Are you sure you want to proceed with this operation? [y/N] " response
+    if [ "${response,,}" != "y" ]; then
+
+        exit 4
+    fi
+    JAR_FIND_SUFFIX="*"
+fi
+
 # If the TARGET_RELEASE hasn't been explicitly specified, try to determine this from github
 if [[ -z "$TARGET_RELEASE" ]]; then
 
     # The value between the dollar-colon tokens is automatically substituted when committing to git.
     # Do not modify this value or the tokens
-    SCRIPT_COMMIT=$(echo "$:d587d4f78979ed63e0e657dc131e486699438052:$" | cut -d '$' -f 2 | cut -d ':' -f 2)
+    SCRIPT_COMMIT=$(echo "$:14fd608f0ac11b27048b07c6b66d6a899934f37c:$" | cut -d '$' -f 2 | cut -d ':' -f 2)
 
     echo ""
     echo "Determining release associated with script: $SCRIPT_COMMIT"
@@ -106,7 +136,6 @@ if [[ -z "$TARGET_RELEASE" ]]; then
     exit 2
 fi
 
-export MATCHED_JAR_FILE_NAME=hadoop-azure
 RELEASE_INFO=$(curl "${GITHUB_API_ROOT_URI}/releases/tags/${TARGET_RELEASE}")
 JAR_ASSET=$(echo $RELEASE_INFO | jq -r '.assets[] | select(.content_type == "application/java-archive") | .')
 if [[ -z "$JAR_ASSET" ]]; then
@@ -140,17 +169,20 @@ fi
 [[ "${DIR_PREFIX}" != */ ]] && DIR_PREFIX="${DIR_PREFIX}/"
 [[ "${HDFS_DIR_PREFIX}" != */ ]] && HDFS_DIR_PREFIX="${HDFS_DIR_PREFIX}/"
 
-if [ -e $LOCAL_PATCH_PATH ]; then 
+if [ $ROLLBACK -eq 0 ]; then
 
-    rm $LOCAL_PATCH_PATH; 
-fi
-echo ""
-echo "Downloading $REMOTE_PATCH_PATH to $LOCAL_PATCH_PATH"
-wget $REMOTE_PATCH_PATH -O $LOCAL_PATCH_PATH
-if [ $? -ne 0 ]; then
+    if [ -e $LOCAL_PATCH_PATH ]; then 
 
-    echo "ERROR: failed to download $REMOTE_PATCH_PATH to $LOCAL_PATCH_PATH"
-    exit 3
+        rm $LOCAL_PATCH_PATH; 
+    fi
+    echo ""
+    echo "Downloading $REMOTE_PATCH_PATH to $LOCAL_PATCH_PATH"
+    wget $REMOTE_PATCH_PATH -O $LOCAL_PATCH_PATH
+    if [ $? -ne 0 ]; then
+
+        echo "ERROR: failed to download $REMOTE_PATCH_PATH to $LOCAL_PATCH_PATH"
+        exit 3
+    fi
 fi
 
 echo ""
@@ -159,45 +191,79 @@ GZs=$(find "$DIR_PREFIX" -name "*.tar.gz" -print0 | xargs -0 zgrep "$MATCHED_JAR
 for GZ in $GZs
 do
 
-    if [[ ! -e "${GZ}.original" ]]; then
+    echo $GZ
+    if [ $ROLLBACK -eq 0 ]; then
 
-        cp "$GZ" "${GZ}.original"
+        if [[ ! -e "${GZ}${BACKUP_SUFFIX}" ]]; then
+
+            cp "$GZ" "${GZ}${BACKUP_SUFFIX}"
+        fi
+
+        ARCHIVE_DIR="${GZ}.dir"
+        if [[ -d $ARCHIVE_DIR ]]; then
+
+            rm -rf "$ARCHIVE_DIR"
+        fi
+        mkdir "$ARCHIVE_DIR"
+        echo "    tar -C "$ARCHIVE_DIR" -zxf $GZ"
+        tar -C "$ARCHIVE_DIR" -zxf "$GZ"
+    else
+
+        # Rollback changes
+        if [[ -e "${GZ}${BACKUP_SUFFIX}" ]]; then
+
+            echo "    cp ${GZ}${BACKUP_SUFFIX} $GZ"
+            cp "${GZ}${BACKUP_SUFFIX}" "$GZ" 
+            rm "${GZ}${BACKUP_SUFFIX}"
+        fi
     fi
-
-    ARCHIVE_DIR="${GZ}.dir"
-    if [[ -d $ARCHIVE_DIR ]]; then
-
-        rm -rf "$ARCHIVE_DIR"
-    fi
-    mkdir "$ARCHIVE_DIR"
-    echo "tar -C "$ARCHIVE_DIR" -zxf $GZ"
-    tar -C "$ARCHIVE_DIR" -zxf "$GZ"
 done
 
 echo ""
-echo "Updating all JAR files with the same name in $DIR_PREFIX$MATCHED_JAR_FILE_NAME*.jar"
-for DST in $(find "$DIR_PREFIX" -name "$MATCHED_JAR_FILE_NAME*.jar" -a ! -name "*datalake*")
+echo "Updating all JAR files with the same name in $DIR_PREFIX$MATCHED_JAR_FILE_NAME*.jar$JAR_FIND_SUFFIX"
+for DST in $(find "$DIR_PREFIX" -name "$MATCHED_JAR_FILE_NAME*.jar$JAR_FIND_SUFFIX" -a ! -name "*datalake*")
 do
     echo $DST
-    # Different handling for symlink or real file
-    if [[ ! -h "$DST" ]]; then
+    if [ $ROLLBACK -eq 0 ]; then
 
-        # Backup original JAR if not already backed up
-        if [[ ! -e "${DST}.original" ]]; then
+        # Backup original file (jar or symlink) if not already backed up
+        if [[ ! -e "${DST}${BACKUP_SUFFIX}" ]]; then
 
-            cp "$DST" "${DST}.original"
+            cp "$DST" "${DST}${BACKUP_SUFFIX}"
         fi
+        # Different handling for symlink or real file
+        if [[ ! -h "$DST" ]]; then
 
-        # Replace with patched JAR
-        rm -f "$DST"
-        DST="$(dirname "$DST")/$PATCHED_JAR_FILE_NAME.jar"
-        echo "cp $LOCAL_PATCH_PATH $DST"
-        cp "$LOCAL_PATCH_PATH" "$DST"
+            # Replace with patched JAR
+            rm -f "$DST"
+            DST="$(dirname "$DST")/$PATCHED_JAR_FILE_NAME.jar"
+            echo "    cp $LOCAL_PATCH_PATH $DST"
+            cp "$LOCAL_PATCH_PATH" "$DST"
+        else
+
+            # For symlink, assume the target will be replaced with the correctly named file. Just update the link.
+            NEW_TARGET="$(dirname $(readlink "$DST"))/$PATCHED_JAR_FILE_NAME.jar"
+            ln -sfn "$NEW_TARGET" "$DST"
+        fi
     else
 
-        # For symlink, assume the target will be replaced with the correctly named file. Just update the link.
-        NEW_TARGET="$(dirname $(readlink "$DST"))/$PATCHED_JAR_FILE_NAME.jar"
-        ln -sfn "$NEW_TARGET" "$DST"
+        # Rollback changes - need to handle 2 cases; hadoop-azure*.jar.original -> hadoop-azure*.jar & hadoop-azure*.jar -> rm
+        DST_FILENAME=$(basename "$DST")
+        DST_EXTENSION=.${DST_FILENAME##*.}
+        if [[ "$DST_EXTENSION" == "$BACKUP_SUFFIX" ]]; then
+
+            # hadoop-azure*.jar.original -> hadoop-azure*.jar
+            DST_ORIG=$(dirname "$DST")/$(basename "$DST" $BACKUP_SUFFIX)
+            echo "    cp ${DST} $DST_ORIG"
+            cp "${DST}" "$DST_ORIG"
+            rm "${DST}"
+
+        elif [[ "$DST_EXTENSION" == "$JAR_EXTENSION" ]]; then
+
+            # hadoop-azure*.jar -> rm
+            echo "    rm $DST"
+            rm "$DST"
+        fi
     fi
 done
 
@@ -205,32 +271,57 @@ done
 if [ $APPLY_HDFS_PATCH -gt 0 ]; then
 
     echo ""
-    echo "Updating all JAR files on HDFS matching; $HDFS_DIR_PREFIX$MATCHED_JAR_FILE_NAME*.jar"
-    for HDST in $(sudo -u $HDFS_USER hadoop fs -find "$HDFS_DIR_PREFIX" -name "$MATCHED_JAR_FILE_NAME*.jar" | grep -v "datalake")
+    echo "Updating all JAR files on HDFS matching; $HDFS_DIR_PREFIX$MATCHED_JAR_FILE_NAME*.jar$JAR_FIND_SUFFIX"
+    for HDST in $(sudo -u $HDFS_USER hadoop fs -find "$HDFS_DIR_PREFIX" -name "$MATCHED_JAR_FILE_NAME*.jar$JAR_FIND_SUFFIX" | grep -v "datalake")
     do
 
-        sudo -u $HDFS_USER hadoop fs -test -e "${HDST}.original"
-        if [ $? -ne 0 ]; then
+        if [ $ROLLBACK -eq 0 ]; then
 
-            sudo -u $HDFS_USER hadoop fs -cp "$HDST" "${HDST}.original"
+            sudo -u $HDFS_USER hadoop fs -test -e "${HDST}${BACKUP_SUFFIX}"
+            if [ $? -ne 0 ]; then
+
+                sudo -u $HDFS_USER hadoop fs -cp "$HDST" "${HDST}${BACKUP_SUFFIX}"
+            fi
+
+            sudo -u $HDFS_USER hadoop fs -rm $HDST
+            HDST="$(dirname "$HDST")/$PATCHED_JAR_FILE_NAME.jar"
+            echo "    hadoop fs -put -f $LOCAL_PATCH_PATH $HDST"
+            sudo -u $HDFS_USER hadoop fs -put -f "$LOCAL_PATCH_PATH" "$HDST"
+        else
+
+            # Rollback changes - need to handle 2 cases; hadoop-azure*.jar.original -> hadoop-azure*.jar & hadoop-azure*.jar -> rm
+            HDST_FILENAME=$(basename "$HDST")
+            HDST_EXTENSION=.${HDST_FILENAME##*.}
+            if [[ "$HDST_EXTENSION" == "$BACKUP_SUFFIX" ]]; then
+
+                # hadoop-azure*.jar.original -> hadoop-azure*.jar
+                HDST_ORIG=$(dirname "$HDST")/$(basename "$HDST" $BACKUP_SUFFIX)
+                echo "    hadoop fs -cp $HDST $HDST_ORIG"
+                sudo -u $HDFS_USER hadoop fs -cp "$HDST" "$HDST_ORIG" 
+                sudo -u $HDFS_USER hadoop fs -rm "$HDST_ORIG"
+
+            elif [[ "$HDST_EXTENSION" == "$JAR_EXTENSION" ]]; then
+
+                # hadoop-azure*.jar -> rm
+                echo "    hadoop fs -rm $HDST
+                sudo -u $HDFS_USER hadoop fs -rm "$HDST"
+            fi
         fi
-
-        sudo -u $HDFS_USER hadoop fs -rm $HDST
-        HDST="$(dirname "$HDST")/$PATCHED_JAR_FILE_NAME.jar"
-        echo "hadoop fs -put -f $LOCAL_PATCH_PATH $HDST"
-        sudo -u $HDFS_USER hadoop fs -put -f "$LOCAL_PATCH_PATH" "$HDST"
     done
 fi
 
-echo ""
-echo "Updating all .tar.gz"
-for GZ in $GZs
-do
+if [ $ROLLBACK -eq 0 ]; then
 
-    echo "tar -czf $GZ -C ${GZ}.dir"
-    tar -czf "$GZ" -C "${GZ}.dir" .
-    rm -rf "${GZ}.dir"
-done
+    echo ""
+    echo "Updating all .tar.gz"
+    for GZ in $GZs
+    do
+
+        echo "    tar -czf $GZ -C ${GZ}.dir"
+        tar -czf "$GZ" -C "${GZ}.dir" .
+        rm -rf "${GZ}.dir"
+    done
+fi
 
 if [ $APPLY_HDFS_PATCH -gt 0 ]; then
 
@@ -240,50 +331,64 @@ if [ $APPLY_HDFS_PATCH -gt 0 ]; then
     do
 
         echo "$HGZ"
-        # Create backup
-        sudo -u $HDFS_USER hadoop fs -test -e "${HGZ}.original"
-        if [ $? -ne 0 ]; then
+        if [ $ROLLBACK -eq 0 ]; then
 
-            sudo -u $HDFS_USER hadoop fs -cp "$HGZ" "${HGZ}.original"
-        fi
+            # Create backup
+            sudo -u $HDFS_USER hadoop fs -test -e "${HGZ}${BACKUP_SUFFIX}"
+            if [ $? -ne 0 ]; then
 
-        # Get the archive, update it with the new jar, repackage the archive & copy it to HDFS
-        ARCHIVE_NAME=$(basename $HGZ)
-        ARCHIVE_DIR=/tmp/${ARCHIVE_NAME}.dir
-        LOCAL_TAR_FILE=/tmp/$ARCHIVE_NAME
-
-        if [[ -e $LOCAL_TAR_FILE ]]; then
-
-            rm -f $LOCAL_TAR_FILE;
-        fi
-        sudo -u $HDFS_USER hadoop fs -copyToLocal "$HGZ" "$LOCAL_TAR_FILE"
-
-        if [[ -d $ARCHIVE_DIR ]]; then
-
-            rm -rf $ARCHIVE_DIR
-        fi
-        mkdir $ARCHIVE_DIR
-        tar -xzf $LOCAL_TAR_FILE -C $ARCHIVE_DIR
-
-        for DST in $(find $ARCHIVE_DIR -name "$MATCHED_JAR_FILE_NAME*.jar" -a ! -name "*datalake*")
-        do
-
-            # Backup original JAR if not already backed up
-            if [[ ! -e "${DST}.original" ]]; then
-
-                cp "$DST" "${DST}.original"
+                sudo -u $HDFS_USER hadoop fs -cp "$HGZ" "${HGZ}${BACKUP_SUFFIX}"
             fi
-            rm -f "$DST"
-            cp "$LOCAL_PATCH_PATH" "$(dirname "$DST")/$PATCHED_JAR_FILE_NAME.jar"
-        done
 
-        cd $ARCHIVE_DIR
-        tar -zcf $LOCAL_TAR_FILE *
-        cd ..
+            # Get the archive, update it with the new jar, repackage the archive & copy it to HDFS
+            ARCHIVE_NAME=$(basename $HGZ)
+            ARCHIVE_DIR=/tmp/${ARCHIVE_NAME}.dir
+            LOCAL_TAR_FILE=/tmp/$ARCHIVE_NAME
 
-        echo "hadoop fs -copyFromLocal -p -f $LOCAL_TAR_FILE $HGZ"
-        sudo -u $HDFS_USER hadoop fs -copyFromLocal -p -f "$LOCAL_TAR_FILE" "$HGZ"
-        rm -rf $ARCHIVE_DIR
-        rm -f $LOCAL_TAR_FILE
+            if [[ -e $LOCAL_TAR_FILE ]]; then
+
+                rm -f $LOCAL_TAR_FILE;
+            fi
+            sudo -u $HDFS_USER hadoop fs -copyToLocal "$HGZ" "$LOCAL_TAR_FILE"
+
+            if [[ -d $ARCHIVE_DIR ]]; then
+
+                rm -rf $ARCHIVE_DIR
+            fi
+            mkdir $ARCHIVE_DIR
+            tar -xzf $LOCAL_TAR_FILE -C $ARCHIVE_DIR
+
+            for DST in $(find $ARCHIVE_DIR -name "$MATCHED_JAR_FILE_NAME*.jar" -a ! -name "*datalake*")
+            do
+
+                # Backup original JAR if not already backed up
+                if [[ ! -e "${DST}${BACKUP_SUFFIX}" ]]; then
+
+                    cp "$DST" "${DST}${BACKUP_SUFFIX}"
+                fi
+                rm -f "$DST"
+                cp "$LOCAL_PATCH_PATH" "$(dirname "$DST")/$PATCHED_JAR_FILE_NAME.jar"
+            done
+
+            cd $ARCHIVE_DIR
+            tar -zcf $LOCAL_TAR_FILE *
+            cd ..
+
+            echo "    hadoop fs -copyFromLocal -p -f $LOCAL_TAR_FILE $HGZ"
+            sudo -u $HDFS_USER hadoop fs -copyFromLocal -p -f "$LOCAL_TAR_FILE" "$HGZ"
+            rm -rf $ARCHIVE_DIR
+            rm -f $LOCAL_TAR_FILE
+        else
+
+            # Rollback changes
+            sudo -u $HDFS_USER hadoop fs -test -e "${HGZ}${BACKUP_SUFFIX}"
+            if [ $? -eq 0 ]; then
+
+                echo "    hadoop fs -cp ${HGZ}${BACKUP_SUFFIX} ${HGZ}"
+                sudo -u $HDFS_USER hadoop fs -cp "${HGZ}${BACKUP_SUFFIX}" "$HGZ" 
+                sudo -u $HDFS_USER hadoop fs -rm "${HGZ}${BACKUP_SUFFIX}"
+            fi
+        fi
     done
 fi
+echo "Finished"
